requirements.txt
SpeechRecognition
pyttsx3
openai
pyaudio
transformers
pytorch